demo_05/requirements.txt
alabaster==0.7.12
astroid==2.3.1
autopep8==1.4.4
Babel==2.7.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.8
imagesize==1.1.0
isort==4.3.21
Jinja2==2.10.3
jwt==0.6.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
paho-mqtt==1.4.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.2
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.6
wrapt==1.11.2
